+++ src/ActionResolver.php
@@ -10,6 +10,7 @@ namespace ZeroBoiler\Events;
 
 use Illuminate\Container\Container;
 use ZeroBoiler\Events\Contracts\Triggerable;
+use ZeroBoiler\Events\Exceptions\ActionResolutionException;
 
 /**
  * Resolves a class FQN string to a Triggerable instance from the container.
@@ -28,19 +29,22 @@ final class ActionResolver
     /**
      * Resolve a class FQN to a Triggerable instance.
      *
-     * @throws \InvalidArgumentException if the class does not exist or does not implement Triggerable
+     * @throws \ZeroBoiler\Events\Exceptions\ActionResolutionException if the class does not exist or does not implement Triggerable
      */
     public function resolve(string $class): Triggerable
     {
         // Check if class exists first
         if (! class_exists($class)) {
-            throw new \InvalidArgumentException("Triggerable class {$class} does not exist");
+            throw new ActionResolutionException($class, 'Class does not exist');
         }
 
         $instance = $this->app->make($class);
 
         if (! $instance instanceof Triggerable) {
-            throw new \InvalidArgumentException("Class {$class} must implement ".Triggerable::class);
+            throw new ActionResolutionException(
+                $class,
+                'Class must implement ' . Triggerable::class,
+            );
         }
 
         return $instance;